packages/api_clients/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "api-clients"
version = "1.1.0"
description = "Unified API client framework for Scopus, Crossref, Gemini image generation, and other scholarly APIs"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Henrik Kragh Sørensen"}
]
keywords = ["api", "scopus", "crossref", "gemini", "image-generation", "scholarly", "research", "rate-limiting", "caching"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pandas>=1.0.0",
    "requests>=2.25.0",
    "PyYAML>=5.4.0",
    "tqdm>=4.60.0",
    "caching",  # Internal package from monorepo
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=3.0",
    "black>=22.0",
    "ruff>=0.0.200",
    "jupyter>=1.0.0",
]

[tool.setuptools]
packages = ["api_clients"]

[tool.setuptools.package-data]
api_clients = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
